+++ src/main/java/ladder/model/result/LadderResult.java
@@ -1,4 +1,4 @@
-package ladder.model;
+package ladder.model.result;
 
 public class LadderResult {
