+++ src/main/java/ladder/view/result/output/FormattableView.java
@@ -1,4 +1,4 @@
-package ladder.view.result;
+package ladder.view.result.output;
 
 import ladder.view.View;
