+++ src/test/java/ladder/model/LadderResultTest.java
@@ -1,5 +1,6 @@
 package ladder.model;
 
+import ladder.model.result.LadderResult;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
