+++ src/test/java/ladder/model/NameTest.java
@@ -1,5 +1,6 @@
 package ladder.model;
 
+import ladder.model.participant.Name;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
